created independent user profile service
