+++ src/main/java/com/revature/services/UserService.java
@@ -12,10 +12,8 @@ import java.util.Optional;
 public class UserService {
 
 	private final UserRepository userRepository;
-	private final UserProfileRepository profileRepository;
 
-	public UserService(UserRepository userRepository, UserProfileRepository profileRepository) {
-		this.profileRepository = profileRepository;
+	public UserService(UserRepository userRepository) {
 		this.userRepository = userRepository;
 	}
 
@@ -30,17 +28,8 @@ public class UserService {
 	public Optional<User> findByCredentials(String email, String password) {
 		return userRepository.findByEmailAndPassword(email, password);
 	}
-
 	public User save(User user) {
 		return userRepository.save(user);
-	}
-
-	public UserProfile saveProfile(UserProfile profile) {
-		return profileRepository.save(profile);
-	}
-
-	public Optional<UserProfile> findProfileForUser(int userId) {
-		return profileRepository.findByUserId(userId);
 	}
 
 	public Optional<User> findByEmail(String email) {
